\name{AdvEYT}
\alias{AdvEYT}
\title{Phenotypic and genotypic information of eleven traits on 190 lines}
\description{
A database comprising 190 lines, each with phenotypic records for eleven traits: `GY_B5IR_F5IR_BEHT`, `GY_BLHT`, `GY_B2IR`, and `GY_TPE` belong to the Grain yield category, `Zinc`, `TKW`, `GRNPRO`, and `ALVW` fall under the Quality family, and `YR_LUD`, `YR_NJ`, and `SR_NJ` correspond to the Diseases category. Furthermore, the dataset includes information on 8,545 SNP-type molecular markers.
}

\usage{
  data(AdvEYT)
}

\details{
  The "AdvEYT" dataset includes the following objects:

  \itemize{
    \item \code{X}: Contains the genotypic data in the form of SNPs molecular markers.
    \item \code{Y}: Contains the phenotypic data of eleven traits.
  }

}