% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApproxMPS.R
\name{ApproxMPS}
\alias{ApproxMPS}
\title{Approximation in Multitrait Parental Selection}
\usage{
ApproxMPS(B0, yHat, R, target = NULL, method = "kl", direction = NULL)
}
\arguments{
\item{B0}{(vector) with the overall mean of each trait of length equal to number of traits (\eqn{t}).}

\item{yHat}{(matrix) of posterior mean of each candidate of dimension \eqn{n \times t}, where
\eqn{n} is the number of individuals in the candidate set.}

\item{R}{(square matrix) of the residual covariance matrix of dimension \eqn{t \times t}.}

\item{target}{(vector) of length equal to number of traits (\eqn{t}) reflecting
the breeder's expectation. Default is NULL.}

\item{method}{(string) the loss function to be used. This must be one of "kl" for Kullback-Leibler, "energy" for Energy Score and "malf" for Multivariate Assymetric Loss. Default is "kl".}

\item{direction}{A vector of length equal to number of traits (\eqn{t}) reflecting the direction
of improvement desired, 1 for increase goal, -1 for decreasing goal.
Default is 1 (increase) for all traits.}
}
\value{
A list with Breeding values, expected loss, and the rank for each candidate of selection.
}
\description{
Approximate the Kullback Leibler, Energy Score or Multivariate Assymetric Loss.
}
\examples{
\dontrun{

}
}
\references{
Villar-Hernández, B.J., et.al. (2018). A Bayesian Decision Theory Approach for Genomic Selection. G3 Genes|Genomes|Genetics. 8(9). 3019–3037
   
   Villar-Hernández, B.J., et.al. (2021). Application of multi-trait Bayesian decision theory for parental genomic selection. 11(2). 1-14
}
