% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AveDistances.R
\name{AveDist}
\alias{AveDist}
\title{Get average distances}
\usage{
AveDist(Xcand, measure = "euclidean")
}
\arguments{
\item{Xcand}{(a matrix) of predictors of the candidates for selection. Dimensions \eqn{n \times k}, where
\eqn{n} is the number of individuals in the candidate set and \eqn{k} is the
number of predictors.}

\item{measure}{(string) the distance measure to be used to calculate average distances between lines. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".}
}
\value{
A vector with average distance for each candidate of selection.
}
\description{
Get average distance for each candidate of selection
}
\examples{
\dontrun{
X <- matrix(rbinom(25, 1, p=0.4), nrow = 5, ncol = 5)
AveDist(X)
}
}
