% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProportion.R
\name{getPropMatrix}
\alias{getPropMatrix}
\title{Get average shared alleles with relatives}
\usage{
getPropMatrix(Markers)
}
\arguments{
\item{Xcand}{(a matrix) of predictors of the candidates for selection. Dimensions \eqn{n \times k}, where
\eqn{n} is the number of individuals in the candidate set and \eqn{k} is the
number of predictors.}
}
\value{
A vector with average shared alleles with relatives for each candidate of selection.
}
\description{
Get average shared alleles with relatives for each candidate of selection
}
