\name{MPS-package}
\alias{MPS-package}
\title{Developed for computing pdf, cdf, quantile, random generation, and estimating the parameters of 24 G-family of statistical distributions.}
\description{Developed for computing the probability density function, computing the cumulative distribution function, computing the quantile function, random generation, and estimating the parameters of 24 G-family of statistical distributions via the maximum product spacing approach introduced in <https://www.jstor.org/stable/2345411>. These families are: beta G distribution due to Eugene et al. (2002), beta exponential G distribution due to Alzaatreh et al. (2013),
beta extended G distribution due to Alzaatreh et al. (2013), exponentiated G distribution due to Gupta et al. (1998),
exponentiated Kumaraswamy G distribution due to Lemonte et al. (2013), exponentiated exponential Poisson G distribution due to Ristic and Nadarajah (2014),
exponentiated generalized G distribution due to Cordeiro et al. (2013), gamma type I G distribution due to Zografos and Balakrishnan (2009),
gamma type II G distribution due to Ristic and Balakrishnan (2012), gamma uniform G distribution due to Torabi and Montazeri (2012),
gamma-X generated of log-logistic family of G distribution due to Alzaatreh et al. (2013), gamma-X family of modified beta exponential G distribution due to Alzaatreh et al. (2013), geometric exponential Poisson G distribution due to Nadarajah et al. (2013), generalized beta G distribution due to Alexander et al. (2012), generalized transmuted G distribution due to Merovci et al. (2017), Kumaraswamy G distribution due to Cordeiro and Castro (2011),
log gamma type I G distribution due to Amini et al. (2013), log gamma type II G distribution due to Amini et al. (2013), Marshall-Olkin G distribution due to Marshall and Olkin (1997), Marshall-Olkin Kumaraswamy G distribution due to Roshini and Thobias (2017), modified beta G distribution due to Nadarajah et al. (2013), odd log-logistic G distribution due to Gauss et al. (2017), truncated-exponential skew-symmetric G distribution due to Nadarajah et al. (2014), and Weibull G distribution due to Alzaatreh et al. (2013).}

\details{
Package: MPS

Type: Package

Version: 2.2.0

Date: 2018-08-25

License: GPL(>=2)
}
