\name{MultiEnvWheat}
\alias{MultiEnvWheat}
\title{Phenotypic and genotypic information of 250 lines evaluated in three environments for two traits}
\description{
A database comprising 250 lines, each with phenotypic records for two traits: `DTHD`, `PTHT`. Each line were evaluated in three environments: `Bed2IR`, `Bed5IR` and `Bed5IR`. Furthermore, the dataset includes the genomic relationship matrix (G). The original data, labeled `DATASET4to5.Wheat.rdata` can be accessed at \url{https://hdl.handle.net/11529/10907}
}

\usage{
  data(MultiEnvWheat)
}

\details{
  The "MultiEnvWheat" dataset includes the following objects:

  \itemize{
    \item \code{G}: Contains the G matrix.
    \item \code{Trigo.Data}: Contains the phenotypic data.
  }

}