\name{ordinalData}
\alias{ordinalData}
\title{Phenotypic and genotypic information (SNPs) of 350 lines with two ordinal traits}
\description{
A database comprising 350 lines, each with phenotypic records for two ordinal traits: `Trait1`, `Trait2`. Furthermore, the dataset includes the genomic data (SNPs) X. This dataset is a toy data to exemplify multitrait ordinal selection.
}

\usage{
  data(ordinalData)
}

\details{
  The "ordinalData" dataset includes the following objects:

  \itemize{
    \item \code{X}: Contains the X (SNPs) matrix.
    \item \code{Y}: Contains the phenotypic data.
  }

}