% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleTraitOPS.R
\name{OrdinalPS}
\alias{OrdinalPS}
\title{Single-Trait Ordinal Parental Selection}
\usage{
OrdinalPS(Xcand, B, thresholds, target = NULL, method = NULL)
}
\arguments{
\item{Xcand}{A matrix of predictors for the candidates under selection, with dimensions \eqn{n \times k}.
\eqn{n}: Number of candidates.
\eqn{k}: Number of predictors.}

\item{B}{A matrix containing regression coefficients, with dimensions \eqn{M \times k}.
\eqn{M}: Number of Markov Chain Monte Carlo (MCMC) samples.
\eqn{k}: Number of predictors.}

\item{thresholds}{A matrix of thresholds values from ordinal regression, with dimensions \eqn{M \times t}, derived from Markov Chain Monte Carlo (MCMC) samples.
Thresholds are the values that define the boundaries between the ordinal categories in the outcome variable. 
They are cut-points on the underlying scale of the latent variable that determine the category boundaries}

\item{target}{A vector that contain probabilities or proportions for the categories within that trait. 
The probabilities/proportions must sum to 1. Zero entries are not allowed.}

\item{method}{A string indicating the loss function to be used. Options are "kl" for Kullback-Leibler, "hellinger" for Hellinger, or "bhattacharyya". Default is "kl".}
}
\value{
A list containing the posterior expected loss, the rank for each candidate, and the probability of each category.
}
\description{
Function to compute Kullback-Leibler, Hellinger, or Bhattacharyya loss in Single-Trait Ordinal Parental Selection.
}
\examples{
\dontrun{
# Example usage of OrdinalMPS
}
}
\references{
Villar-Hernández, B.J., et.al. (2018). A Bayesian Decision Theory Approach for Genomic Selection. G3 Genes|Genomes|Genetics. 8(9). 3019–3037
   
   Villar-Hernández, B.J., et.al. (2021). Application of multi-trait Bayesian decision theory for parental genomic selection. G3 Genes|Genomes|Genetics. 11(2). 1-14
}
