% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimMatrix.R
\name{SimMatrix}
\alias{SimMatrix}
\title{Get average shared alleles with relatives}
\usage{
SimMatrix(Xcand)
}
\arguments{
\item{Xcand}{(a matrix) of predictors of the candidates for selection. Dimensions \eqn{n \times k}, where
\eqn{n} is the number of individuals in the candidate set and \eqn{k} is the
number of predictors.}
}
\value{
A squared matrix of dimension \eqn{n \times n} with proportion of shared alleles with relatives for each candidate of selection.
}
\description{
Get average shared alleles of each candidate with the rest of candidates
}
