% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApproxMPG.R
\name{ApproxMPS}
\alias{ApproxMPS}
\title{Multitrait Parental Selection}
\usage{
ApproxMPS(B0, ETA, R, target, p, method = "kl")
}
\arguments{
\item{B0}{(vector) with the overall mean of each trait of length equal to number of traits (\eqn{t}).}

\item{ETA}{(matrix) of posterior mean of each candidate of dimension \eqn{n \times k}, where
\eqn{n} is the number of individuals in the candidate set.}

\item{R}{(square matrix) of the residual covariance matrix of dimension \eqn{t \times t}.}

\item{p}{(escalar) the proportion of selected candidates.}

\item{method}{(string) indicate the loss function to use. Posible values are "kl" for Kullback-Leibler, "energy" for Energy Score and "malf" for Multivariate Assymetric Loss. Default is "kl".}
}
\value{
A list of BVs, approximated expected loss and a logical vector that indicate what lines are selected.
}
\description{
Approximate the Kullback Leibler, Energy Score or Multivariate Assymetric Loss.
}
