% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastMPS.R
\name{FastMPS}
\alias{FastMPS}
\title{Multitrait Parental Selection}
\usage{
FastMPS(Xcand, B0, B, R, p, method = "kl")
}
\arguments{
\item{Xcand}{(a matrix) of predictors of candidates for selection of dimensions \eqn{n \times k}, where
\eqn{n} is the number of individuals in the candidate set and \eqn{k} is the
number of predictors.}

\item{B0}{(matrix) of dimension \eqn{M \times t} for the intercept term in
linear model, where \eqn{M} is the number of markov chain monte carlo samples and \eqn{t} is the number of traits.}

\item{B}{(array) containing regression coeficients of dimension, \eqn{M \times k \times t} of Markov Chain Monte Carlo samples.}

\item{R}{(matrix) of dimension \eqn{M \times (t \times (t + 1) / 2)} of Markov Chain Monte Carlo samples of the variance-covariance
components in the residual covariance matrix.}

\item{p}{(escalar) the proportion of selected candidates.}

\item{method}{(string) indicate the loss function to use. Posible values are "kl" for Kullback-Leibler, "energy" for Energy Score and "malf" for Multivariate Assymetric Loss. Default is "kl".}
}
\value{
A list of BVs, posterior expected loss and a logical vector that indicate what lines are selected.
}
\description{
Parallelized multiLoss function to compute Kullback Leibler, Energy Score or Multivariate Assymetric Loss in Multitrait Parental Selection.
}
\examples{
\dontrun{
# Clean and setting local environment
rm(list = ls())
setwd("Put your working directory here")
# Loading needed packages
library(scoringGS)
library(BGLR)
# Loading dataset
data(wheat)
Y <- as.matrix(wheat.Y)
X <- scale(wheat.X, center=TRUE, scale = TRUE)
n <- nrow(Y)
# Just for example: parental population
porc_parental <- 0.4
idPar <- sample(1:n, ceiling(porc_parental*n), replace = FALSE)
XTrn <- X[-idPar,]
YTrn <- Y[-idPar,]
# ModelFit using BGLR
ETA <- list(list(X = XTrn, model = "BRR", saveEffects = TRUE))
model <- Multitrait(y = YTrn,
                  ETA = ETA,
                  intercept = TRUE,
                  resCov = list(df0 = 5, S0 = NULL,type = "UN", saveEffects = TRUE),
                  saveAt = paste0("./chains/", "BRR_"),
                  nIter = 50000,
                 burnIn = 20000)
# Reading Posterior MCMC
mu <- as.matrix(read.table(file = "./chains/BRR_mu.dat", header = FALSE))
B <- readBinMatMultitrait('./chains/BRR_ETA_1_beta.bin')
R <- as.matrix(read.table(file = "./chains/BRR_R.dat", header = FALSE))
XPar <- X[idPar, ]
# Evaluating Loss Function
out <- FastMPS(Xcand = XPar,
               B0 = mu,
               B = B,
               R = R,
               p = 0.1,
               method = "kl")
# Plotting results
colnames(out$yHat) <- colnames(Y)
pairs(out$yHat,
     col = ifelse(out$selected, "#2c7fb8", "darkgray"),
     pch = ifelse(out$selected, 19, 1))
}
}
