% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastMPS.R
\name{FastMPS}
\alias{FastMPS}
\title{Fast Multitrait Parental Selection}
\usage{
FastMPS(Xcand, B0, B, R, target = NULL, method = "kl")
}
\arguments{
\item{Xcand}{(a matrix) of predictors of the candidates for selection. Dimensions \eqn{n \times k}, where
\eqn{n} is the number of individuals in the candidate set and \eqn{k} is the
number of predictors.}

\item{B0}{(matrix) of dimension \eqn{M \times t} for the intercept term in the
linear model, where \eqn{M} is the number of markov chain monte carlo samples and \eqn{t} is the number of traits.}

\item{B}{(array) containing regression coeficients of dimension \eqn{M \times k \times t} of Markov Chain Monte Carlo samples.}

\item{R}{(matrix) of dimension \eqn{M \times (t \times (t + 1) / 2)} of Markov Chain Monte Carlo samples of the variance-covariance
components in the residual covariance matrix.}

\item{target}{(vector) of length equal to number of traits (\eqn{t}) reflecting
the breeder's expectation. Default is NULL.}

\item{method}{(string) the loss function to be used. This must be one of "kl" for Kullback-Leibler, "energy" for Energy Score and "malf" for Multivariate Assymetric Loss. Default is "kl".}
}
\value{
A list with Breeding values, posterior expected loss, and the rank for each candidate of selection.
}
\description{
Parallelized function for computing Kullback-Leibler, Energy Score, or Multivariate Asymmetric Loss in Multitrait Parental Selection.
}
\examples{
\dontrun{

}
}
\references{
Villar-Hernández, B.J., et.al. (2018). A Bayesian Decision Theory Approach for Genomic Selection. G3 Genes|Genomes|Genetics. 8(9). 3019–3037
   
   Villar-Hernández, B.J., et.al. (2021). Application of multi-trait Bayesian decision theory for parental genomic selection. 11(2). 1-14
}
