% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimilarity.R
\name{getSim}
\alias{getSim}
\title{Get average similarities}
\usage{
getSim(K)
}
\arguments{
\item{K}{The pedigree or the genomic relationship matrix. Dimensions \eqn{n \times n}, where
\eqn{n} is the number of individuals in the candidate set.}
}
\value{
A vector with average similarities for each candidate of selection.
}
\description{
Get average similarities for each candidate of selection
}
