% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringGS-package.R
\docType{package}
\name{MPS-package}
\alias{MPS}
\alias{MPS-package}
\title{MPS: Multi-trait Parental Selection}
\description{
An R Package to Assist Multi-trait Parental Selection via Genomic, Pedigree, or Phenotypic Selection.
}
\keyword{internal}
