% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringGS-package.R
\docType{package}
\name{MPS-package}
\alias{MPS}
\alias{MPS-package}
\title{MPS: Multi-trait Parental Selection}
\description{
Plant and animal breeders are interested in selecting the best individuals as parents in succesive selection cycles. `MPS` compute Kullback-Leibler (kl), Energy Score (energy) and Multivariate Asymmetric (malf) distances for every potential parent in a population to assist in multitrait selection.
}
\keyword{internal}
