\name{MPS-package}
\alias{MPS-package}
\title{Developed for computing pdf, cdf, quantile, random generation, and estimating the parameters of 24 G-family of statistical distributions.}
\description{Developed for computing the probability density function, computing the cumulative distribution function, computing the quantile function, random generation, and estimating the parameters of 24 G-family of statistical distributions via the maximum product spacing approach introduced in <https://www.jstor.org/stable/2345411>. These families are: beta G distribution due to Eugene et al. (2002), beta exponential G distribution due to Alzaatreh et al. (2013),
beta extended G distribution due to Alzaatreh et al. (2013), exponentiated G distribution due to Gupta et al. (1998),
exponentiated Kumaraswamy G distribution due to Lemonte et al. (2013), exponentiated exponential Poisson G distribution due to Ristic and Nadarajah (2014),
exponentiated generalized G distribution due to Cordeiro et al. (2013), gamma type I G distribution due to Zografos and Balakrishnan (2009),
gamma type II G distribution due to Ristic and Balakrishnan (2012), gamma uniform G distribution due to Torabi and Montazeri (2012),
gamma-X generated of log-logistic family of G distribution due to Alzaatreh et al. (2013), gamma-X family of modified beta exponential G distribution due to Alzaatreh et al. (2013), geometric exponential Poisson G distribution due to Nadarajah et al. (2013), generalized beta G distribution due to Alexander et al. (2012), generalized transmuted G distribution due to Merovci et al. (2017), Kumaraswamy G distribution due to Cordeiro and Castro (2011),
log gamma type I G distribution due to Amini et al. (2013), log gamma type II G distribution due to Amini et al. (2013), Marshall-Olkin G distribution due to Marshall and Olkin (1997), Marshall-Olkin Kumaraswamy G distribution due to Roshini and Thobias (2017), modified beta G distribution due to Nadarajah et al. (2013), odd log-logistic G distribution due to Gauss et al. (2017), truncated-exponential skew-symmetric G distribution due to Nadarajah et al. (2014), and Weibull G distribution due to Alzaatreh et al. (2013).}

\details{
Package: MPS

Type: Package

Version: 2.2.0

Date: 2018-08-25

License: GPL(>=2)
}
\author{
Mahdi Teimouri

Maintainer: Mahdi Teimouri <teimouri@aut.ac.ir>
}

\references{
Alexander, C., Cordeiro, G. M., and Ortega, E. M. M. (2012). Generalized beta-generated distributions, \emph{Computational Statistics and Data Analysis}, 56, 1880-1897.

Alzaatreh, A., Lee, C., and Famoye, F. (2013). A new method for generating families of continuous distributions, \emph{Metron}, 71, 63-79.

Amini, M., MirMostafaee, S. M. T. K., and Ahmadi, J. (2013). Log-gamma-generated families of distributions, \emph{Statistics}, 48 (4), 913-932.

Cheng, R. C. H. and Stephens, M. A. (1989). A goodness-of-fit test using Moran's statistic with estimated parameters, \emph{Biometrika}, 76 (2), 385-392.

Cordeiro, G. M. and  Castro, M. (2011). A new family of generalized distributions, \emph{Journal of Statistical Computation and Simulation}, 81, 883-898.

Cordeiro, G. M., Ortega, E. M. M., and  da Cunha, D. C. C. (2013). The exponentiated generalized class of distributions, \emph{Journal of Data Science}, 11, 1-27.

Eugene, N., Lee, C., and Famoye, F. (2002). Beta-normal distribution and its applications, \emph{Communications in Statistics-Theory and Methods}, 31, 497-512.

Gupta, R. C., Gupta, P. L., and Gupta, R. D. (1998). Modeling failure time data by Lehman alternatives, \emph{Communications in Statistics-Theory and Methods}, 27, 887-904.

Gauss, M. C., Alizadeh, M., Ozel, G., Hosseini, B. Ortega, E. M. M., and Altunc, E. (2017). The generalized odd log-logistic family of distributions: properties, regression models and applications, \emph{Journal of Statistical Computation and Simulation}, 87(5), 908-932.

Lemonte, A. J.,  Barreto-Souza, W., and  Cordeiro, G. M. (2013). The exponentiated Kumaraswamy distribution and its log-transform, \emph{Brazilian Journal of Probability and Statistics}, 27, 31-53.

Marshall, A. W. and Olkin, I. (1997). A new method for adding a parameter to a family of distributions with application to the exponential and Weibull families, \emph{Biometrika}, 84, 641-652.

Merovcia, F., Alizadeh, M., Yousof, H. M., and Hamedani, G. G. (2017). The exponentiated transmuted-G family of distributions: Theory and applications, \emph{Communications in Statistics-Theory and Methods}, 46(21), 10800-10822.

Nadarajah, S., Cancho, V. G., and Ortega, E. M. M. (2013). The geometric exponential Poisson distribution, \emph{Statistical Methods & Applications}, 22, 355-380.

Nadarajah, S., Teimouri, M., and Shih, S. H. (2014). Modified beta distributions, \emph{Sankhya}, 76 (1), 19-48.

Nadarajah, S., Nassiri, V., and Mohammadpour, A. (2014). Truncated-exponential skew-symmetric distributions, \emph{Statistics}, 48 (4), 872-895.

Ristic, M. M. and Balakrishnan, N. (2012). The gamma exponentiated exponential distribution, \emph{Journal of Statistical Computation and Simulation}, 82, 1191-1206.

Ristic, M. M. and Nadarajah, S. (2014). A new lifetime distribution, \emph{Journal of Statistical Computation and Simulation}, 84 (1), 135-150.

Roshini, G. and Thobias, S. (2017). Marshall-Olkin Kumaraswamy Distribution, \emph{International Mathematical Forum},  12 (2), 47-69.

Torabi, H. and Montazeri, N. H. (2012).  The gamma uniform distribution and its applications, \emph{Kybernetika}, 48, 16-30.

Zografos, K. and Balakrishnan, N. (2009). On families of beta- and generalized gamma-generated distributions and associated inference, \emph{Statistical Methodology}, 6, 344-362.}
